+++ js/devel/0-core.js
@@ -32,7 +32,7 @@
                 }
             },
             entityURLs: {
-                /* "entityCode": value, etc. */
+                /* "entityCode": value, etc. See docs. */
             },
             _requestDelay: 0
         };
@@ -678,12 +678,61 @@
         mydata.preloaderElement.append(elements);
     };
 
+    var _simpleDialog = function(opts) {
+        var options = $.extend({}, {
+            content: null,
+            width: "auto"
+        }, opts);
+        var deferred = $.Deferred();
+        var closeLink = $("<a class='ua-close-link'><span class='ua-icon ua-icon-16 ua-icon-close'></span></a>");
+        var div = $("<div>")
+            .css("position", "relative")
+            .append(closeLink)
+            .append(options.content);
+        div.dialog({
+            dialogClass: "ua-panic-dialog",
+            resizable: false,
+            modal: true,
+            width: options.width,
+            height: "auto",
+            open: function(event, ui) {
+                $('.ui-widget-overlay').on('click', function() {
+                    $(this).siblings('.ui-dialog').find('.ui-dialog-content').dialog('close');
+                });
+            },
+            closeText: $.usosCore.lang("Zamknij", "Close"),
+            close: function() {
+                deferred.resolve();
+            },
+            show: {
+                effect: 'fade',
+                duration: 300
+            },
+            hide: {
+                effect: 'fade',
+                duration: 100
+            }
+        });
+        closeLink.on("click", function() {
+            div.dialog("close");
+        });
+
+        return deferred.promise();
+    };
+
     $[NS] = {
+
+        /* Undocumented methods. These don't have to stay backward-compatible
+         * (and may be removed). */
+
         _getSettings: function() { return mydata.settings; },
         _console: fixedConsole,
         _methodForwarder: _methodForwarder,
         _preload: _preload,
         _nlang: _nlang,
+        _simpleDialog: _simpleDialog,
+
+        /* Documented methods. */
 
         init: init,
         usosapiFetch: usosapiFetch,